% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{auc_eer_opt.t_generation}
\alias{auc_eer_opt.t_generation}
\title{Calculate AUC, EER, optimal threshold for ROC curves}
\usage{
auc_eer_opt.t_generation(ROC_list)
}
\arguments{
\item{ROC_list}{List of sensitivity and sepecificity from many different methods}
}
\description{
With the list type of data containing sensitivity and sepecificity values from many thresholds, it calcuates AUC, EER, optiimal threshold
to give the minumum summation of FPR and FNR.
}
