% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boosted_clique}
\alias{boosted_clique}
\title{Function to perform boosted clique on two input circles}
\usage{
boosted_clique(circle_in, circle_ref, ncross_in_bins = 30, xbins_in = 20,
  ncross_in_bin_size = 1, ncross_ref_bins = NULL, xbins_ref = 30,
  ncross_ref_bin_size = NULL, eps = 0.75, seed = 1, num_cores = 8,
  plot = TRUE, verbose = FALSE, cl = NULL)
}
\arguments{
\item{circle_in}{The input circle}

\item{circle_ref}{The reference circle}

\item{ncross_in_bins}{Number of bins in the input circle (See smart_sample)}

\item{xbins_in}{Number of bins along each axis in the hexbin grid for the input circle}

\item{ncross_in_bin_size}{Number of points to sample from each bin in the input circle}

\item{ncross_ref_bins}{Number of bins in the reference circle}

\item{xbins_ref}{Number of bins along each axis in the hexbin grid for the reference circle}

\item{ncross_ref_bin_size}{Number of points to sample from each bin in the reference circle}

\item{eps}{Distance tolerance for declaring an edge match}

\item{seed}{The random seed for reproducing results}

\item{num_cores}{The number of processor cores for parallel processing}

\item{plot}{If TRUE, produce a plot of the clique results}

\item{verbose}{If TRUE, print out the timing results for each portion of the algorithm}

\item{cl}{Optionally, a parallel cluster that has already been created}
}
\value{
The statistics for the matching between the two circles
}
\description{
Function to perform boosted clique on two input circles
}
