% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edge_fix_circle_mat}
\alias{edge_fix_circle_mat}
\title{Match input and reference with given circle information in input. Circle in the reference will be also fixed.}
\usage{
edge_fix_circle_mat(input, referece, input_circles, ref_circles)
}
\arguments{
\item{input}{Input image for fixing three circles}

\item{input_circles}{Three circles which will be fixed in the input}

\item{ref_circles}{Three circles which will be fixed in the reference too.}

\item{reference}{Input image for finding correspondence}
}
\description{
Match between circle 1 in the input impression vs. circle 1 in the referece impression. Repeat this for circle 2 and circle 3.
This matching function is for comparing degraded impressions to see how worse the method perform when comparing similar regions.
}
