% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{focus_data2}
\alias{focus_data2}
\title{Cut first 1% and 99% quantile on x and y coordinates}
\usage{
focus_data2(data)
}
\arguments{
\item{data}{matrix of entire points from shoe impression}
}
\description{
Function to cut first 1% and 99% quantile on x and y coordinates
}
