% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gg_circle}
\alias{gg_circle}
\title{Adding the circlular area}
\usage{
gg_circle(r, xc, yc, color = "black", fill = NA, ...)
}
\arguments{
\item{r}{radius of the circle}

\item{xc}{x value of the circle}

\item{yc}{y value of the circle}
}
\description{
Function to find the circular area
}
