% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initial_circle}
\alias{initial_circle}
\title{Find the initial circle position by quantile of x and y ranges}
\usage{
initial_circle(input)
}
\arguments{
\item{input}{input impression}
}
\description{
Function to find the initial circle position by quantile of x and y ranges. Centers are found as (30%, 80%), (20%, 40%), (70%, 70%) with radius of 50 for three circles.
}
