% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initial_circle2}
\alias{initial_circle2}
\title{Find three circles}
\usage{
initial_circle2(input)
}
\arguments{
\item{input}{Questioned impression}
}
\description{
Find three circles with different positions for degraded impressions
}
