% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{input_example}
\alias{input_example}
\title{An example of an input shoeprint}
\format{a data frame}
\usage{
input_example
}
\description{
An example of an input shoeprint
}
\keyword{datasets}
