% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{int_inside_center}
\alias{int_inside_center}
\title{Detecting points within the circle}
\usage{
int_inside_center(Data, r, nseg, c1, c2)
}
\arguments{
\item{Data}{The input print}

\item{r}{radius of the circle}

\item{c1}{x value of the circle}

\item{c2}{y value of the circle}
}
\description{
Function to find points within the circle
}
