% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{leftcorner_cent}
\alias{leftcorner_cent}
\title{Adjust it to (0,0) in the left lower corner}
\usage{
leftcorner_cent(data)
}
\arguments{
\item{data}{The input data}
}
\description{
Adjust it to (0,0) in the left lower corner
}
