% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_print}
\alias{match_print}
\title{Matches an input and a reference shoeprint}
\usage{
match_print(print_in, print_ref, circles_input = NULL,
  circles_reference = NULL, ncross_in_bins = 30, xbins_in = 20,
  ncross_in_bin_size = 1, ncross_ref_bins = NULL, xbins_ref = 30,
  ncross_ref_bin_size = NULL, max_rotation_angle = 60, eps = 0.75,
  seed = 1, num_cores = 8, plot = FALSE, verbose = FALSE)
}
\arguments{
\item{print_in}{The input print}

\item{print_ref}{The reference print}

\item{circles_input}{The input circles, specified as a matrix with columns (x, y, rad), or NULL to automatically generate}

\item{circles_reference}{The reference circles, specified as a matrix with columns (x, y, rad), or NULL to automatically generate}

\item{ncross_in_bins}{Number of bins in the input circle (See smart_sample)}

\item{xbins_in}{Number of bins along each axis in the hexbin grid for the input circle}

\item{ncross_in_bin_size}{Number of points to sample from each bin in the input circle}

\item{ncross_ref_bins}{Number of bins in the reference circle}

\item{xbins_ref}{Number of bins along each axis in the hexbin grid for the reference circle}

\item{ncross_ref_bin_size}{Number of points to sample from each bin in the reference circle}

\item{max_rotation_angle}{The maximum rotation angle, in degrees, for inclusion in the best circle matches}

\item{eps}{Distance tolerance for declaring an edge match}

\item{seed}{The random seed for reproducing results}

\item{num_cores}{The number of processor cores for parallel processing}

\item{plot}{If TRUE, produce a plot of the clique results}

\item{verbose}{If TRUE, print out the timing results for each portion of the algorithm}
}
\value{
The statistics for the matching between the two circles
}
\description{
Function to perform matching of footprints by matching all the candidate circles with input circles. This function performs initial matching between 3 circles on the input print and 27 candidate circles on the reference shoeprint. After the initial cliques, it select 2 best circles for each input circle and performs reinforcement matching on them. In reinforcement matching, full points on reference circle are considered instead of user defined. The function returns the final circle parameters and the statistics of the triangle formed.
}
\examples{
\dontrun{
data(input_example)
data(reference_example)

## Transform all the points to have (0,0) at lower left corner.
print_in <- leftcorner_cent(reference_example)
print_ref <- leftcorner_cent(reference_example)

## Perform Print Match
print_stats <- match_print(print_in, print_ref,
                          ncross_in_bins = 30, xbins_in = 20, ncross_in_bin_size = 1,
                          ncross_ref_bins = NULL, xbins_ref = 30, ncross_ref_bin_size = NULL,
                          eps = .75, seed = 1, num_cores = parallel::detectCores(),
                          plot = TRUE, verbose = FALSE)
print_stats
}
}
