% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_print_subarea}
\alias{match_print_subarea}
\title{Match input and reference with given circle information in input. The searching area is confined.}
\usage{
match_print_subarea(input, reference, input_circles, max_rotation_angle)
}
\arguments{
\item{input}{Input image for fixing three circles}

\item{reference}{Input image for finding correspondence}

\item{input_circles}{Three circles which will be fixed in the input}

\item{max_rotation_angle}{maximum rotation angle that we allow for comparisons.}
}
\description{
Find corresponding areas in reference for fixed circles in input. To reduce the time, it confines the area for candidate circles in reference image.
For the first circle, it confines the area into left toe, for the second circle, it searches only left bottom, for the third circle, it searches right toe area of reference shoe.
}
