% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{reference_example}
\alias{reference_example}
\title{An example of a reference shoeprint}
\format{a data frame}
\usage{
reference_example
}
\description{
An example of a reference shoeprint
}
\keyword{datasets}
