% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rev_R}
\alias{rev_R}
\title{Reverse right side shoe to look like left side shoe}
\usage{
rev_R(Data)
}
\arguments{
\item{Data}{data with x and y coordinate values from right side of shoe}
}
\description{
Function to transform coordinate values to look like from the left side shoe
}
