% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simple_auc}
\alias{simple_auc}
\title{Calculate AUC}
\usage{
simple_auc(TPR, FPR)
}
\arguments{
\item{TPR}{True positive rate}

\item{FPR}{False positive rate}
}
\description{
Calculate AUC with information of TPR and FPR
}
