% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_plot}
\alias{start_plot}
\title{Starting plot to position three local areas of shoe Q}
\usage{
start_plot(input, reference, input_circle)
}
\arguments{
\item{input}{The input print (shoe Q)}

\item{reference}{The reference print (shoe K)}

\item{input_circle}{Centers and radius for three circles that we fix in the input print (shoe Q)}
}
\description{
Function to plot shoe Q and shoe K together. In shoe Q, three circles are colored.
}
