% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{step0_plot}
\alias{step0_plot}
\title{First plot before starting matching to confirm the circle position.}
\usage{
step0_plot(input, reference, input_circles)
}
\arguments{
\item{input}{Input image for fixing three circles}

\item{reference}{Input image for finding correspondence}

\item{input_circles}{Three circles which will be fixed in the input}
}
\description{
Draw edge coordiates with designated circles. Circle will be colored in both input and reference
}
