% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{step0_Q_K_plot}
\alias{step0_Q_K_plot}
\title{Draw plot of two shoe impressions with each designated circles}
\usage{
step0_Q_K_plot(input, reference, circle_in, circle_ref)
}
\arguments{
\item{input}{Questioned impression (Q)}

\item{reference}{Known impression (K)}

\item{circle_in}{Circle information for Q}

\item{circle_ref}{Circle information for K}
}
\description{
Draw plot of questioned and reference shoe with each designated circles.
}
