% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_list_match}
\alias{summarize_list_match}
\title{Summarize MC-COMP macthing across three circles}
\usage{
summarize_list_match(list_mat, mat_class, size)
}
\arguments{
\item{list_mat}{Matching result as list type}

\item{mat_class}{Class either KM or KNM}

\item{size}{the size of the shoe}
}
\description{
Get average of five similarity features and standard deviaitions of rotation angle estimations across three circle matching.
}
