% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{focus_data2}
\alias{focus_data2}
\title{Cut x and y coordinate points less than 1% quantile and larger than 99% quantile}
\usage{
focus_data2(data)
}
\arguments{
\item{data}{data with x and y coordinate values}
}
\description{
Function to cut x and y coordinate points less than 1% quantile and larger than 99% quantile
}
